Add blog post for recsys 2024
Would like workshop papers to be linked (don't have a preference if it's directly to the published work, or to the post about the paper on the lab website).

Also I think Molly has a picture of us at the lab dinner, waiting to hear back from them and then will add it in!